Crate trust_graph

Modules

certificate_serde

Structs

Certificate
Chain of trusts started from self-signed root trust.

KeyPair
An Ed25519 keypair.

PublicKeyHashable
Wrapper to use PublicKey in HashMap

Trust
One element in chain of trust in a certificate. TODO delete pk from Trust (it is already in a trust node)

TrustGraph
Graph to efficiently calculate weights of certificates and get chains of certificates. TODO serialization/deserialization TODO export a certificate from graph

Functions

current_time